requirements.txt
openai==0.27.8
anthropic==0.2.10
typer[all]==0.9.0
langchain==0.0.205
load-dotenv==0.1.0
python-decouple==3.8.0
